+++ app/Http/Controllers/UsuarioController.php
@@ -30,7 +30,7 @@ class UsuarioController extends Controller
 
     public function create()
     {
-        $status=$this->status->all()->where('id','<>',2)->sortby('status');
+        $status=$this->status->all()->sortby('status');
         $funcionarios=$this->funcionarios->all()->where('id','<>',1);
         return view('usuarios.novo', compact('status', 'funcionarios'));
     }
@@ -67,38 +67,58 @@ class UsuarioController extends Controller
         ]);
     }
 
-    public function show($id)
-    {
-        //
-    }
-
     public function edit($id)
     {
         $usuario=$this->usuarios->find($id);
+
+        if(!$usuario){
+            return redirect('usuarios')->with('warning', 'Usuário não encontrado!');
+        }
+
         $status=$this->status->all()->where('id','<>',2)->sortby('status');
         $funcionarios=$this->funcionarios->all()->where('id','<>',1);
+
         return view('usuarios.editar', compact('usuario', 'status', 'funcionarios'));
     }
 
     public function update(Request $request, $id)
     {
+
+        $usuario = $this->usuarios->find($id);
+        
+        if(!$usuario){
+            return json_encode([
+                'success'=> false,
+                'msg' => 'Usuário não encontrado!',
+            ]);
+        }
+        
         $this->validate($request,[
             'login' => 'required|unique:usuarios,username,'.$id,
-            'senha' => 'required',
-            'confirmacao' => 'required|same:senha',
+            'senha' => session('user.tipo') == 0 ? 'required' : '',
+            'confirmacao' => session('user.tipo') == 0 ? 'required|same:senha' : '',
             'status' => 'required|numeric|exists:tipo_status,id',
-            'tipo' => 'required|in:0,1',
+            'tipo' => session('user.tipo') == 1 ? 'required|in:0,1' : '',
             'funcionario' => 'required|exists:funcionarios,id|unique:usuarios,funcionario_id,'.$id
         ]);
 
         try{
-            $this->usuarios->find($id)->update([
-                'username' => $request->login,
-                'password' => bcrypt($request->senha),
-                'tipo' => $request->tipo,
-                'tipo_status_id' => $request->status,
-                'funcionario_id' => $request->funcionario
-            ]);
+            if(session('user.tipo') == 1){
+                $this->usuarios->find($id)->update([
+                    'username' => $request->login,
+                    'tipo' => $request->tipo,
+                    'tipo_status_id' => $request->status,
+                    'funcionario_id' => $request->funcionario
+                ]);
+            }else{
+                $this->usuarios->find($id)->update([
+                    'username' => $request->login,
+                    'password' => bcrypt($request->senha),
+                    'tipo_status_id' => $request->status,
+                    'funcionario_id' => $request->funcionario
+                ]);
+            }
+            
         }catch(QueryException $ex){ 
             return json_encode([
                 'success'=> false,
@@ -114,6 +134,13 @@ class UsuarioController extends Controller
 
     public function destroy($id)
     {
+
+        $usuario = $this->usuarios->find($id);
+        
+        if(!$usuario){
+            return back()->with('warning', 'Usuário não encontrado!');
+        }
+
         try {             
             $this->usuarios->destroy($id);
         }catch(QueryException $ex){
